apps/server/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "amish-pratap-singh-1",email = "amishpratapsingh13@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "pydantic-settings",
    "psycopg2-binary",
    "python-jose",
    "bcrypt",
    "python-multipart",
    "pydantic[email] (>=2.12.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]